Only allow the creator of the event to delete it

+++ commands/events/delete_event.py
@@ -5,6 +5,7 @@ def delete_event(update, context):
     chat_id = update.effective_chat.id
     event_args = context.args
     db_connector, events_col = db_handler()
+    message = None
     if len(event_args) != 1:
         message = f"Invalid message, the only parameter should be the event name."
         return send_message(context, chat_id, message)
@@ -12,9 +13,15 @@ def delete_event(update, context):
     event_name = event_args[0]
 
     query = {'name': event_name.lower(), "chat_id": chat_id}
-    deleted = db_connector.delete_record(events_col, query)
-    if deleted == 0:
+    event = db_connector.search_record(events_col, query)
+    if not event:
         message = f"Event {event_name} not found."
-        return send_message(context, chat_id, message)
-    message = f"Event {event_name} deleted."
+    if not message and event['creator']['id'] == update.message.from_user.id:
+        deleted = db_connector.delete_record(events_col, query)
+        if deleted == 0:
+            message = f"There was an error deleting the event."
+        else:
+            message = f"Event {event_name} deleted."
+    if not message:
+        message = "Only the creator of the event can delete it."
     return send_message(context, chat_id, message)
